First commited as before_18886_rev1

+++ lttng/org.eclipse.linuxtools.ctf.core/src/org/eclipse/linuxtools/ctf/core/event/types/Definition.java
@@ -15,7 +15,7 @@ package org.eclipse.linuxtools.ctf.core.event.types;
 import org.eclipse.linuxtools.ctf.core.event.io.BitBuffer;
 
 /**
- * A CTF definition
+ * A CTF definiton
  *
  * A definition is like an object of a declaration class. It fills the declaration
  * with values. <br>
@@ -137,8 +137,8 @@ public abstract class Definition {
      * @since 2.2
      */
     protected static void alignRead(BitBuffer input, IDeclaration declaration){
-        long align = declaration.getAlignment();
-        long pos = input.position() + ((align - (input.position() % align)) % align);
+        int align = (int) declaration.getAlignment();
+        int pos = input.position() + ((align - (input.position() % align)) % align);
         input.position(pos);
     }
